+++ target/modest_client/handlers.h
@@ -22,6 +22,8 @@ ETERM *
 handle_insert_before(ErlMessage* emsg, ETERM *response);
 ETERM *
 handle_insert_after(ErlMessage* emsg, ETERM *response);
+ETERM *
+handle_replace(ErlMessage* emsg, ETERM *response);
 
 typedef struct _state_t {
   int fd;